Add z-rotation tracking (#4)
Added new PiBot functionality for tracking the robot's rotation using the gyroscope. A prototype implementation of this was tested as functioning.

To start, invoke _gyro_start() on the bot object. To stop, invoke _gyro_stop(). While the gyro is running, all you have to do is check the _rotation_z variable of the robot. No other function calls are necessary.